Fix: breakdown into subcomponents per file
